added generic unix and windows shell files
