+++ apps/platforms/linux/kde-konsole/kde-konsole.py
@@ -10,40 +10,6 @@ app.name: konsole
 
 @ctx.action_class("user")
 class user_actions:
-    # terminal-tag functions implementation
-    def terminal_list_directories():
-        actions.insert("ls")
-        actions.key("enter")
-
-    def terminal_list_all_directories():
-        actions.insert("ls -a")
-        actions.key("enter")
-
-    def terminal_change_directory(path: str):
-        actions.insert("cd {}".format(path))
-        if path:
-            actions.key("enter")
-
-    def terminal_change_directory_root():
-        actions.insert("cd /")
-        actions.key("enter")
-
-    def terminal_clear_screen():
-        actions.insert("clear")
-        actions.key("enter")
-
-    def terminal_run_last():
-        actions.key("up enter")
-
-    def terminal_kill_all():
-        actions.key("ctrl-c")
-        actions.insert("y")
-        actions.key("enter")
-
-    def terminal_rerun_search(command: str):
-        actions.key("ctrl-r")
-        actions.insert(command)
-
 
     # tabs-tag functions implementations
     def tab_jump(number):
