+++ apps/platforms/linux/kde-konsole/kde-konsole.talon
@@ -1,8 +1,12 @@
 os: linux
 and app.name: konsole
 -
-tag(): terminal 
 # now activates generic_terminal
+tag(): terminal 
+
+# activates the implementation of the function of the terminal tag
+tag(): user.generic_unix_shell
+
 tag(): user.git
 tag(): user.tabs
 tag(): user.anaconda
