+++ apps/platforms/win/powershell/power_shell.py
@@ -65,40 +65,3 @@ class UserActions:
     def file_manager_open_volume(volume: str):
         """file_manager_open_volume"""
         actions.user.file_manager_open_directory(volume)
-
-    # generic_terminal function implementations
-
-    def terminal_list_directories():
-        actions.insert("ls")
-        actions.key("enter")
-
-    def terminal_list_all_directories():
-        actions.insert("ls -force")
-        actions.key("enter")
-
-    def terminal_change_directory(path: str):
-        actions.insert("cd {}".format(path))
-        if path:
-            actions.key("enter")
-
-    def terminal_change_directory_root():
-        """Root of current drive"""
-        actions.insert("cd /")
-        actions.key("enter")
-
-    def terminal_clear_screen():
-        """Clear screen"""
-        actions.insert("cls")
-        actions.key("enter")
-
-    def terminal_run_last():
-        actions.key("up enter")
-
-    def terminal_rerun_search(command: str):
-        actions.key("ctrl-r")
-        actions.insert(command)
-
-    def terminal_kill_all():
-        actions.key("ctrl-c")
-        actions.insert("y")
-        actions.key("enter")
